requirements.txt
Cython==0.19.1
Django==1.5.1
GDAL==1.9.1
PySAL==1.6.0
South==0.8.1
argparse==1.2.1
distribute==0.6.24
django-dajax==0.9.2
django-dajaxice==0.5.5
django-extensions==1.1.1
django-model-utils==1.4.0
django-pandas==0.1.0
django-tastypie==0.9.15
emcee==1.2.0
gunicorn==17.5
ipython==0.13.2
matplotlib==1.2.1
mimeparse==0.1.3
nose==1.3.0
numpy==1.7.1
pandas==0.12.0
psycopg2==2.5.1
python-dateutil==2.1
python-nvd3==0.7.0
pytz==2013b
scikit-learn==0.13.1
scikits.sparse==0.1-dev
scipy==0.12.0
six==1.3.0
wsgiref==0.1.2
yolk==0.4.3
